requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # multipart/form-data uploads

# ── Config & validation ───────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.1.0

# ── Google Gemini ─────────────────────────────────────────────────────────────
google-generativeai==0.5.4

# ── Google Cloud Vision (OCR) ─────────────────────────────────────────────────
google-cloud-vision==3.7.2

# ── Google Drive ─────────────────────────────────────────────────────────────
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
aiohttp==3.9.5                   # async HTTP for Drive API calls

# ── Qdrant ────────────────────────────────────────────────────────────────────
qdrant-client==1.9.1

# ── Neo4j ─────────────────────────────────────────────────────────────────────
neo4j==5.20.0

# ── Redis ─────────────────────────────────────────────────────────────────────
redis==5.0.4                     # standard redis-py with asyncio support

# ── Token counting ────────────────────────────────────────────────────────────
tiktoken==0.7.0

# ── Async utilities ───────────────────────────────────────────────────────────
anyio==4.3.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                    # AsyncClient for FastAPI test client
